Getting reducers working
Working on question 3, question 4 is next.

+++ question-two/src/main/java/com/revature/driver/Driver.java
@@ -3,7 +3,7 @@ package com.revature.driver;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
@@ -34,7 +34,7 @@ public class Driver extends Configured implements Tool {
 		job.setReducerClass(Reduce.class);
 
 		job.setOutputKeyClass(Text.class);
-		job.setOutputValueClass(IntWritable.class);
+		job.setOutputValueClass(DoubleWritable.class);
 
 		boolean success = job.waitForCompletion(true);
 		return success ? 0 : 1;
